Fix stuck client when socket client never initializes
There was an issue with the Client implementation: if a socket started a connection on the internal MJPEG port and was never initialized,
the entire server thread would be stuck in the while loop waiting for initialization.
This broke other clients connected to the server: no frames would be received anymore.

To reproduce on emulator / real device:
- start an Appium server with mJPEG server enabled (`appium:mjpegServerPort:9100` for instance)
- Start an Appium session that reads the mJPEG frames (using Appium Inspector for instance)
- Then from `adb shell`: connect to the internal mJPEG server (default port 7810):
`nc localhost 7810`
- --> Streaming will be broken for the existing clients

This fixes the problem by returning early: if the client is not ready initialization will be retried later if necessary (or never) instead of blocking the thread

Another side effect is that we actually receive frames from `nc localhost 7810` instead of it being stuck